# documentation/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "description": "",
        "title": "Road Protect API",
        "version": "1.4.1"
    },
    "paths": {
...
        "/api/v1/query/vehicle": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "mine",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "enum": [
                            "all",
                            "owned",
                            "using",
                            "infringementProjectionsContracts",
                            "infringementProjectionsInfringements",
                            "infringementProjectionsWithoutInfringements",
                            "infringementProjectionsNoContracts"
                        ],
                        "in": "query",
                        "name": "via",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "default": "false",
                        "enum": [
                            "true",
                            "false"
                        ],
                        "in": "query",
                        "name": "graphing",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "enum": [
                            "owner",
                            "user",
                            "hybrid"
                        ],
                        "in": "query",
                        "name": "accountRole",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "in": "query",
                        "name": "startDate",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "in": "query",
                        "name": "endDate",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "items": {
                                "$ref": "#/definitions/Vehicle"
                            },
                            "type": "array"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Invalid query parameters"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Vehicles"
                ],
                "operationId": "VehicleQueryController_getMany",
                "summary": "Query and filter vehicles"
            }
        },
        "/api/v1/query/vehicle/spreadsheet": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "mine",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "enum": [
                            "all",
                            "owned",
                            "using",
                            "infringementProjectionsContracts",
                            "infringementProjectionsInfringements",
                            "infringementProjectionsWithoutInfringements",
                            "infringementProjectionsNoContracts"
                        ],
                        "in": "query",
                        "name": "via",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "default": "false",
                        "enum": [
                            "true",
                            "false"
                        ],
                        "in": "query",
                        "name": "graphing",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "enum": [
                            "owner",
                            "user",
                            "hybrid"
                        ],
                        "in": "query",
                        "name": "accountRole",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "in": "query",
                        "name": "startDate",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "in": "query",
                        "name": "endDate",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "items": {
                                "$ref": "#/definitions/AdvancedTableColumn"
                            },
                            "type": "array"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "",
                        "schema": {
                            "type": "object",
                            "title":"Buffer"
                        }
                    },
                    "400": {
                        "description": "E112: No data to export to spreadsheet"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Invalid query parameters"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Vehicles"
                ],
                "operationId": "VehicleQueryController_getManyAsSpreadsheet",
                "summary": "Query and filter vehicles as a spreadsheet"
            }
        },
...
    },
    "definitions": {
...
        "AdvancedTableColumn": {
            "title":"AdvancedTableColumn",
            "properties": {
                "isDisplaying": {
                    "default": true,
                    "type": "boolean"
                },
                "key": {
                    "type": "string"
                },
                "spreadsheetFormat": {
                    "default": "default",
                    "enum": [
                        "currency",
                        "date",
                        "number",
                        "default"
                    ],
                    "type": "string"
                },
                "spreadsheetKey": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            },
            "required": [
                "key",
                "title",
                "isDisplaying",
                "spreadsheetKey",
                "spreadsheetFormat"
            ],
            "type": "object"
        },
...
        "Vehicle": {
            "title": "Vehicle",
            "properties": {
                "autoAssignTo": {
                    "enum": [
                        "Owner",
                        "User"
                    ],
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "color": {
                    "type": "string"
                },
                "contracts": {
                    "description": "Contract[]",
                    "items": {
                        "type": "string"
                    },
                    "type": "array"
                },
                "createdAt": {
                    "type": "string"
                },
                "currentLeaseContract": {
                    "description": "LeaseContract",
                    "type": "object"
                },
                "currentOwnershipContract": {
                    "description": "OwnershipContract",
                    "type": "object"
                },
                "infringements": {
                    "description": "Infringement",
                    "type": "object"
                },
                "locationRecords": {
                    "description": "IturanLocationRecord[]",
                    "type": "object"
                },
                "logs": {
                    "description": "Log[]",
                    "type": "object"
                },
                "manufacturer": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "modelYear": {
                    "type": "string"
                },
                "registration": {
                    "type": "string"
                },
                "type": {
                    "default": "Private",
                    "enum": [
                        "Private",
                        "Truck"
                    ],
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "vehicleId": {
                    "type": "number"
                },
                "weight": {
                    "type": "number"
                }
            },
            "required": [
                "createdAt",
                "updatedAt",
                "vehicleId",
                "registration",
                "manufacturer",
                "model",
                "modelYear",
                "color",
                "category",
                "weight",
                "type",
                "contracts",
                "currentLeaseContract",
                "currentOwnershipContract",
                "infringements",
                "autoAssignTo",
                "logs",
                "locationRecords"
            ],
            "type": "object"
        }
    },
    "securityDefinitions": {
        "bearer": {
            "in": "header",
            "name": "Authorization",
            "type": "apiKey"
        }
    },
...
}